% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_df_term.R
\name{class_df_from_term}
\alias{class_df_from_term}
\title{Title}
\usage{
class_df_from_term(model, class_post_formula = FALSE)
}
\arguments{
\item{model}{an 'lm' object}

\item{class_post_formula}{if FALSE (default) returns class of df col BEFORE any in-formula transformations applied to terms
if TRUE returns class of df col AFTER any in-formula transformations applied to terms}
}
\value{
a named list containing named character vectors. the list names are model terms. the vector names are data frame column names. the vector values are characters for the data frame column's class
}
\description{
Title
}
