% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{The clean_name() function}
\usage{
clean_name(name_unique)
}
\arguments{
\item{name_unique}{a character vector containing a set of names to be parsed}
}
\value{
a character vector that has symbols removed from the supplied input 'name_unique'
}
\description{
The clean_name() function
}
\examples{
# not run
# library(dplyr)
# name_all = c('joseph','joe','jo ',' joesephine','joseph','jo')
# name_all \%>\% clean_name()
#
# name_all_unique = name_all \%>\% clean_name() \%>\% unique \%>\% sort
#
# name_all_unique \%>\%
#   # other spot check manual gsubs
#   gsub(.,pattern='n/a',replace='') \%>\%
#   gsub(.,pattern='N/A',replace='') \%>\%
#   gsub(.,pattern="can't answer.*",replace='') \%>\%
#   gsub(.,pattern='dk',replace='') \%>\%
#   gsub(.,pattern='^x$',replace='')
}
