% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_tzen.R
\name{colname_zen}
\alias{colname_zen}
\title{the colname_zen() function zenifies a data frame's column names by lower and caterpillar casing}
\usage{
colname_zen(df, suffix = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{suffix}{a single character string to attach as a suffix to column names}
}
\value{
a data.frame whose column names have been lower and caterpillar cased
}
\description{
the colname_zen() function zenifies a data frame's column names by lower and caterpillar casing
}
\examples{
iris \%>\% colname_zen() \%>\% head

iris \%>\% colname_zen(.,suffix="src_1") \%>\% head
}
