% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datzen-package.R
\docType{package}
\name{datzen}
\alias{datzen}
\alias{datzen-package}
\title{datzen: A package containing miscellaneous helper functions the Zen way.}
\description{
The datzen package provides three categories of commonly used Zen functions:
data in out, themes (rebranded functions and visual), and special computation functions.
The goal of this package is to help data analysts rapidly prototype common tasks.
}
\section{datzen functions}{

To contribute or see function internals, see
 \url{https://github.com/mikejacktzen/datzen}
}

