% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yankovise.R
\name{dictate_outin}
\alias{dictate_outin}
\title{The dictate_outin() function}
\usage{
dictate_outin(swap_out, swap_in)
}
\arguments{
\item{swap_out}{vector whose elements will be used as the names of the resulting character vector}

\item{swap_in}{vector whose elements will be used as the values of the resulting character vector}
}
\value{
a character vector with values from swap_in and names from swap_out
}
\description{
An extremely simple function that creates a named vector.
A useful helper to create key-value dictionaries to use in \code{\link[stringr]{str_replace_all}}
}
\examples{
dictate_outin('compton','bompton')

}
\seealso{
\code{\link[stringr]{str_replace_all}}
}
