% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garble.R
\name{garble}
\alias{garble}
\title{the garble function to spit out random gibberish}
\usage{
garble(size_out, pattern_bank = "~!@#$\%^&*()-+")
}
\arguments{
\item{size_out}{an integer for length of output, must be 3 or larger}

\item{pattern_bank}{a single string of patterns (no space)}
}
\value{
a character vector of random characters
}
\description{
the garble function to spit out random gibberish
}
\examples{

garble(1)
garble(10)
garble(10,'!&@')
garble(3)
}
