% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebranding.R
\name{\%nin\%}
\alias{\%nin\%}
\title{The "Not In" Operator \%nin\%, negation of \code{\link[base]{match}}}
\usage{
... \%nin\% NA
}
\arguments{
\item{...}{see \%in\% details in \code{\link[base]{match}}}
}
\value{
a logical vector
}
\examples{
c(1,2,3) \%nin\% c(1,2)
c("a","b") \%nin\% c("a","b")
}
\seealso{
\code{\link[base]{match}}
}
