% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterload.R
\name{iterload}
\alias{iterload}
\title{The iterload() helper function to easily read in .rds files output from itersave()}
\usage{
iterload(dir_rds)
}
\arguments{
\item{dir_rds}{path of directory containing .rds files}
}
\value{
a list resulting from lapply(.,FUN=readRDS)
}
\description{
The iterload() helper function to easily read in .rds files output from itersave()
}
