% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itersave.R
\name{itersave}
\alias{itersave}
\title{The itersave() function to iteratively run-then-write the output of a user supplied function}
\usage{
itersave(func_user, vec_arg_func, mainDir, subDir, subSubDir = "/failed/",
  start = 1, end = length(vec_arg_func), parallel = FALSE)
}
\arguments{
\item{func_user}{a user supplied function taking in a simple argument of one-element}

\item{vec_arg_func}{a vector whose elements will be iteratively used as arguments in func_user}

\item{mainDir}{parent directory to write output to}

\item{subDir}{sub directory to write good results to}

\item{subSubDir}{sub sub directory to write bad results to}

\item{start}{integer to start iteration. Defaults to 1}

\item{end}{integer to end iteration. Defaults to length(func_user)}

\item{parallel}{logical to use doParallel backend. Defaults to FALSE}
}
\description{
The itersave() function to iteratively run-then-write the output of a user supplied function
}
