% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cache.R
\name{load_cache}
\alias{load_cache}
\title{The load_cache() function to load .rdata files from a knitr cache}
\usage{
load_cache(dir_cache)
}
\arguments{
\item{dir_cache}{a single character string for the directory containing the .rdata files}
}
\value{
NULL, but the .rdata files are loaded into the global environment
}
\description{
The load_cache() function to load .rdata files from a knitr cache
}
\examples{
dir_cache = here::here("/writeup/article_tech_doc_cache/html/")
load_cache(dir_cache)
}
