% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebranding.R
\name{logit}
\alias{logit}
\title{The logit() function is a rebranding of \code{\link[stats]{qlogis}}}
\usage{
logit(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{...}{see args of \code{\link[stats]{qlogis}}}
}
\value{
a numeric
}
\examples{
plot(x=seq(from=0,to=1,by=0.01),y=logit(seq(from=0,to=1,by=0.01)))
identical(logit(0.5),stats::qlogis(0.5))
}
\seealso{
\code{\link[stats]{qlogis}}
}
