% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_op.R
\name{moran_op}
\alias{moran_op}
\title{The moran_op() function}
\usage{
moran_op(x_features, coord_nb)
}
\arguments{
\item{x_features}{a data.frame of features that will be coerced as.matrix()}

\item{coord_nb}{a data.frame of the nb-coordinate triplets (i,j,k) returned from ?save_coord_nb()}
}
\value{
A list of quantities related to the Moran Operator.
NOTE: the list structure is ready to be passed to the STAN mcmc software
}
\description{
The moran_op() function
}
\examples{
library(spdep); example(columbus)
x_features = columbus$AREA
coord_nb = save_coord_nb(foo_spdf=columbus)
moran_op(x_features,coord_nb)
}
