% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2html_all.R
\name{rd2html_all}
\alias{rd2html_all}
\title{The rd2html_all() function to batch convert rd files to html files}
\usage{
rd2html_all(dir_man_rd, dir_man_html)
}
\arguments{
\item{dir_man_rd}{a character string for the directory containing .rd files}

\item{dir_man_html}{a character string for the directory to output .html files}
}
\value{
nothing is returned, but external .html files are written to dir_man_html
}
\description{
uses the tools::Rd2HTML function inside base::lapply
}
\examples{

}
\seealso{
\code{\link[tools]{Rd2HTML}}
}
