% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd2md_all.R
\name{rd2md_all}
\alias{rd2md_all}
\title{The rd2md_all() function batch converts .rd files to .md files}
\usage{
rd2md_all(dir_man_rd, dir_man_md)
}
\arguments{
\item{dir_man_rd}{a character string for the directory containing .rd files}

\item{dir_man_md}{a character string for the directory to output .md files}
}
\value{
nothing is returned, but external .md files are written to dir_man_md
}
\description{
uses the Rd2md::Rd2markdown function inside base::lapply
}
\examples{

}
\seealso{
\code{\link[Rd2md]{Rd2markdown}}
}
