% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readss.R
\name{readss}
\alias{readss}
\title{The readss() function reads a csv, but returns a sampled subset of rows}
\usage{
readss(ss = 10, replace = TRUE, ...)
}
\arguments{
\item{ss}{an integer for desired row sample size}

\item{replace}{a logical picking with/without-replacement sampling}

\item{...}{args passed to \code{\link{read.csv}}}
}
\value{
a 'data.frame' with subsampled rows
}
\description{
The readss() function reads a csv, but returns a sampled subset of rows
}
\examples{
test1 <- c(1:5, "6,7", "8,9,10")
tf <- tempfile()
writeLines(test1, tf)

read.csv(tf, fill = TRUE) # 1 column
readss(ss = 5,tf,replace=TRUE)
readss(ss = 100,tf,replace=FALSE)
readss(ss = 5,tf,replace=FALSE)
}
\seealso{
\code{\link[utils]{read.csv}}
}
