% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_allbut.R
\name{rm_allbut}
\alias{rm_allbut}
\title{rm_allbut}
\usage{
rm_allbut(list_keep = NULL, envir = globalenv())
}
\arguments{
\item{list_keep}{an optional list of names of objects to keep from '?rm()'.
The default 'list_keep=NULL' will remove all objects.}

\item{envir}{an optional environment. Defaults to '?globalenv()'.}
}
\value{
nothing but a message will be printed for the kept objects
}
\description{
rm_allbut
}
