% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_df_schema.R
\name{save_df_schema}
\alias{save_df_schema}
\title{The save_df_schema() function}
\usage{
save_df_schema(x, outdir = getwd(), ...)
}
\arguments{
\item{x}{a data.frame whose metadata will be used as a validating schema}

\item{outdir}{a character designating the output filepath}

\item{...}{optional arguments}
}
\value{
NULL but a physical .json file is written to disk
}
\description{
The save_df_schema() function
}
\examples{
iris2=cbind(iris,Species_chr=as.character(iris$Species))
save_df_schema(x=iris2)
jsonlite::read_json('~/projects/datzen/tests/iris2_schema.json',simplifyVector = TRUE)
}
