% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_df_schema.R
\name{scheme}
\alias{scheme}
\title{The scheme() function creates a schema for a data.frame}
\usage{
scheme(x, outdir = NULL, pattern_null = NULL, pattern_sep_row = NULL,
  pattern_sep_col = NULL, pattern_quote_string = NULL,
  pattern_escape_quote = NULL, description_col = NA, ...)
}
\arguments{
\item{x}{a data.frame whose properties will be schemed() to a schema.}

\item{outdir}{a character string designating the output filepath (default NULL).}

\item{pattern_null}{a character string that contains the pattern for NULL values (default NULL).}

\item{pattern_sep_row}{a character string that contains the pattern for seperating rows (default NULL).}

\item{pattern_sep_col}{a character string that contains the pattern for seperating columns (default NULL).}

\item{pattern_quote_string}{a character string that contains the pattern for quoting character values (default NULL).}

\item{pattern_escape_quote}{a character string that contains the pattern for escaping quotes (default NULL).}

\item{description_col}{a named character vector whose name elements are the column names of data.frame x and
each value element is a brief human-readable description of the column of x (default NA).}

\item{...}{optional arguments}
}
\value{
A nested list with 'schema_global' and 'schema_local' .
Note if the input argument 'outdir' is not NULL, a physical .json file is written to disk at the specified directory.
}
\description{
The schema should describe global characteristics of the data.frame (eg the number of rows and columns)
and local characteristics of the data.frame (eg each column name and type).
The resulting schema is saved as an external json file that complements the csv version of data.frame x.
}
\details{
This assumes the values in data.frame x contain the exhaustive scoped-population values
NOTE: Does not have to be population, just ensure coverage / support.
}
\examples{
iris2=cbind(iris,Species_chr=as.character(iris$Species))
scheme(x=iris2,outdir='~/projects/datzen/tests/')
jsonlite::read_json('~/projects/datzen/tests/iris2_schema.json',simplifyVector = TRUE)
}
\seealso{
\code{\link[datzen]{mandate}}
}
