% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlm.R
\name{simlm}
\alias{simlm}
\title{The simlm function}
\usage{
simlm(n = 10, p = 4, coef_true = seq(1:p), seed = NULL)
}
\arguments{
\item{n}{integer for number of observations}

\item{p}{integer for number of cols in design matrix}

\item{coef_true}{vector (length p) for true lm coefficients}

\item{seed}{a seed for RNG (default NULL)}
}
\value{
a 3 element list with coef_true, yx, noise.

The column names of yx are names(yx)=c('y',paste0("x",seq(1:p)))
}
\description{
Quickly simulate data from a basic linear model.

Each of the n noise elements are from \code{\link[stats]{rnorm}}(mean=0,sd=1).

Each of the n x p design-matrix elements are from \code{\link[stats]{runif}}(min=0,max=1).

The outcome y is noise added to the linear combination of the design matrix with 'coef_true'.
}
\examples{

simlm()

out = simlm(n=1000,p=10)
lm(data=out$yx,y~-1+.)

out = simlm(n=1000,seed=123)
lm(data=out$yx,y~-1+.)
}
