% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_tab.R
\name{tree_tab}
\alias{tree_tab}
\title{The tree_tab() function to conveniently tabulate hierarchical counts}
\usage{
tree_tab(formula_tab, dat_in)
}
\arguments{
\item{formula_tab}{a formula}

\item{dat_in}{a dataframe of columns usable by formula}
}
\value{
a list of tables and a tree
}
\description{
The tree_tab() function to conveniently tabulate hierarchical counts
}
\examples{
names(starwars)
test = tree_tab(1 ~ gender + eye_color + skin_color +
                  mass + height,starwars)
# View(test)

test$list_tab_count[[1]]
test$list_tab_count[[1]] \%>\% tally(nn)  # removed NAs
test$list_tab_count[[5]]

test$list_tab_count[[5]] \%>\% tally(n)  # counted NAs
nrow(starwars)

test$tree_print

}
