% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yankocise.R
\name{yankocise}
\alias{yankocise}
\title{The yankocise() function to Wierd Al Yankocise a string}
\usage{
yankocise(snip, brev = TRUE, dict_out_in = NULL, suffix = NULL)
}
\arguments{
\item{snip}{a character string of lyrics}

\item{brev}{a logical (default TRUE) determining whether to pass 'snip' into brevitweet(snip) 
which abbreviates 'snip' for twitter brevity}

\item{dict_out_in}{a named character vector (default NULL) where an element's name will later be 
swapped out for its associated value. When NULL, an internally hardcoded 'dict_out_in' is used. 
See \code{\link[stringr]{str_replace_all}}}

\item{suffix}{a character string used as a suffix, say for a signed name.}
}
\value{
a character string that has been Weird Al yankocised() using elements
in 'dict_out_in' swapping value 4 name.
}
\description{
The yankocise() function to Wierd Al Yankocise a string
}
\examples{

snip = "Suede sun roof, hanging out the big top We leave the dealership, head to the rim shop"

yankocise(snip,brev=FALSE)
yankocise(snip)
yankocise(snip,brev=TRUE,suffix="- @2chainz")

# user supplied dictionary
dict_out_in = c(dictate_out_in('dealership','SERVER ROOM'),
                dictate_out_in('big','LAP'),
                dictate_out_in('rim','RAM'))

yankocise(snip,brev=TRUE,suffix="- @2chainz",dict_out_in = dict_out_in)
}
