% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_zen.R
\name{colname_zen}
\alias{colname_zen}
\title{The colname_zen() function zenifies a data frame's column names by lower and caterpillar casing}
\usage{
colname_zen(df, suffix = NULL, ind_suffix = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{suffix}{a single character string (optional) to attach as a suffix to column names (default NULL)}

\item{ind_suffix}{an integer vector (optional) that picks the specific columns to attach the suffix (default NULL)}
}
\value{
a modified character vector of names that have been lower and caterpillar cased with an optional suffix
}
\description{
The colname_zen() function zenifies a data frame's column names by lower and caterpillar casing
}
\examples{
df_foo = iris

df_foo \%>\% colname_zen()
df_foo \%>\% colname_zen(.,suffix="src_1")
df_foo \%>\% colname_zen(.,suffix="src_1",c(1,5))

# assign names to original data frame
names_new = df_foo \%>\% colname_zen(.,suffix="src_1",c(1,5))
names(df_foo) = names_new
names(df_foo)
}
