% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebranding.R
\name{expit}
\alias{expit}
\title{The expit() function is a rebranding of \code{\link[stats]{plogis}}}
\usage{
expit(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{...}{see args of see args of \code{\link[stats]{plogis}}}
}
\value{
a numeric
}
\examples{
plot(x=seq(from=-10,to=10,by=1),y=expit(seq(from=-10,to=10,by=1)))
identical(expit(10),stats::plogis(10))
}
\seealso{
\code{\link[stats]{plogis}}
}
