% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadss.R
\name{freadss}
\alias{freadss}
\title{The freadss() function reads in a csv and perhaps subsets the rows (optionally sampled).}
\usage{
freadss(input, ss = NULL, replace = TRUE, ind_choose = NULL, ...)
}
\arguments{
\item{input}{character for file path of csv passed to \code{\link[data.table]{fread}}}

\item{ss}{integer for desired row sample size. Default of ss is NULL, meaning no subsampling.}

\item{replace}{a logical picking with/without-replacement sampling}

\item{ind_choose}{optional integer vector of specific rows to read in (instead of sampling)}

\item{...}{optional args passed to \code{\link[data.table]{fread}}}
}
\value{
a 'data.frame' with optionally subsetted rows (perhaps from sampling)
}
\description{
The input file is read in-memory via \code{\link[data.table]{fread}}.
If rows are subset, there is a speed down. Hence, subsetting rows takes on a one-time slow down
in order to bean count the read-in memory footprint
}
\examples{

set.seed(1)
m = matrix(rnorm(1e5),ncol=10)
csv = data.frame(x=1:1e4,m)

setwd("~/projects/datzen/tests/proto/")
write.csv(csv,"test.csv")

dir_test = '~/projects/datzen/tests/proto/test.csv'

result = freadss(input=dir_test,ss=1000,replace=TRUE,ind_choose=NULL)
# sometimes sample drops some row

dim(result)
head(result)

ind_pick = as.integer(50:100)
result = freadss(input=dir_test,replace=TRUE,
                 ind_choose=ind_pick,
                 drop='V12')

identical(sort(as.integer(result$V1)),sort(ind_pick))

dim(result)
head(result)
tail(result)

# no sub sampling
result = freadss(input=dir_test,ss=NULL)
dim(result)
}
\seealso{
\code{\link[data.table]{fread}}
}
