% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itersave.R
\name{itersave}
\alias{itersave}
\title{The itersave() function to iteratively run-then-write (to .rds) the output of a user supplied function}
\usage{
itersave(func_user, vec_arg_func, mainDir, subDir, subSubDir = "/failed/",
  beg = 1, end = length(vec_arg_func), parallel = FALSE)
}
\arguments{
\item{func_user}{a user supplied function taking in a simple argument of one-element}

\item{vec_arg_func}{a named vector whose elements' value will be iteratively used as arguments in func_user.
Further, the vector elements' name will be used as the .rds filename.}

\item{mainDir}{parent directory to write output to}

\item{subDir}{sub directory to write good results to}

\item{subSubDir}{sub sub directory to write bad results to}

\item{beg}{integer to begin iteration. Defaults to 1}

\item{end}{integer to end iteration. Defaults to length(vec_arg_func)}

\item{parallel}{logical to use library(doParallel) backend. Defaults to FALSE}
}
\description{
The user supplied function, in the argument 'func_user', is expected to return an object that itersave() can further save physically.
So, itersave() will write a physical .rds file to the user specified directory.
The name of the physical file will use the elements' names in the named vector 'vec_arg_func' as the filename. eg if names(vec_arg_func[[i]])='foo', then the physical file is 'foo.rds'
Note: itersave() uses purrr::safely() with if-else to handle errors.
}
\examples{

foo_func_spec = function(x){return(log(x))}

mainDir = '~/projects/datzen/tests/proto/temp/'
subDir = '/dump_1/'
subSubDir = '/failed/'

arg_vec_spec = 1:10
# using arg vector index as suffix
names(arg_vec_spec) = paste0('arg_foo',seq_along(arg_vec_spec))
# using last 6 digits of arg value as suffix
names(arg_vec_spec) = paste0('arg_foo',stringr::str_sub(arg_vec_spec,start=-6))

itersave(func_user=foo_func_spec,vec_arg_func=arg_vec_spec,
         mainDir,subDir,subSubDir='/failed/',parallel=FALSE)

list.files('~/projects/datzen/tests/proto/temp/dump_1/')
out=readRDS('~/projects/datzen/tests/proto/temp/dump_1/10.rds')
identical(out,log(10))
out=readRDS('~/projects/datzen/tests/proto/temp/dump_1/1.rds')
identical(out,log(1))

# error control
arg_vec_spec = 'a'
itersave(func_user=foo_func_spec,vec_arg_func=arg_vec_spec,
         mainDir,subDir,subSubDir='/failed/',parallel=FALSE)

out=readRDS('~/projects/datzen/tests/proto/temp/dump_1/failed/a.rds')
}
