% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebranding.R
\name{logit}
\alias{logit}
\title{the logit() function, rebranding of stats::qlogis}
\usage{
logit(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{...}{see args of \code{\link{qlogis}}}
}
\value{
a numeric
}
\description{
the logit() function, rebranding of stats::qlogis
}
\examples{
plot(x=seq(from=0,to=1,by=0.01),y=logit(seq(from=0,to=1,by=0.01)))
}
\seealso{
\code{\link[stats]{qlogis}}
}
