% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_op.R
\name{moran_op}
\alias{moran_op}
\title{The moran_op() function computes the moran operator}
\usage{
moran_op(x_features, coord_nb)
}
\arguments{
\item{x_features}{a data.frame of features that will be coerced \code{\link[base]{as.matrix}}}

\item{coord_nb}{a data.frame of the nb-coordinate triplets (i,j,k) returned
from \code{\link[datzen]{coordinate_nb}} or \code{\link[spdep]{listw2sn}}}
}
\value{
A list of quantities related to the Moran Operator.
NOTE: the list structure is ready to be passed to the STAN mcmc software
}
\examples{
library(spdep); example(columbus)
x_features = columbus$AREA
coord_nb = save_coord_nb(foo_spdf=columbus)
moran_op(x_features,coord_nb)
}
