% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readss.R
\name{readss}
\alias{readss}
\title{The readss() function reads a csv, but returns a sampled subset of rows}
\usage{
readss(ss = 10, replace = TRUE, ...)
}
\arguments{
\item{ss}{an integer for desired row sample size}

\item{replace}{a logical picking with/without-replacement sampling}

\item{...}{args passed to ?read.csv()}
}
\value{
a 'data.frame' with subsampled rows
}
\description{
The readss() function reads a csv, but returns a sampled subset of rows
}
