% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zen.R
\name{theme_zen}
\alias{theme_zen}
\title{The zen theme for ggplot2}
\usage{
theme_zen()
}
\description{
The zen theme for ggplot2
}
\examples{
qplot(mpg, wt, data = mtcars)
qplot(mpg, wt, data = mtcars) + theme_zen()
}
