% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yankovise.R
\name{yankovise}
\alias{yankovise}
\title{The yankovise() function to Weird Al Yankovise a string}
\usage{
yankovise(snip, brev = TRUE, dict_outin = NULL, suffix = NULL)
}
\arguments{
\item{snip}{a character string of lyrics}

\item{brev}{a logical (default TRUE) determining whether to pass 'snip' into brevitweet(snip)
which abbreviates 'snip' for twitter brevity}

\item{dict_outin}{a named character vector (default NULL) where an element's name will later be
swapped out for its associated value. When NULL, an internally hardcoded 'dict_outin' is used.
See \code{\link[stringr]{str_replace_all}}}

\item{suffix}{a character string used as a suffix, say for a signed name.}
}
\value{
a character string that has been Weird Al yankovised() using elements
in 'dict_outin' swapping value 4 name.
}
\description{
The function helps the user revise a lyric by substituting original words for thematic keywords.
NOTE: the input string will immediately be lowercased via tolower(snip)
}
\examples{

snip = "Suede sun roof, hanging out the big top We leave the dealership, head to the rim shop"

yankovise(snip,brev=FALSE)
yankovise(snip)
yankovise(snip,brev=TRUE,suffix="- @2chainz")

# user supplied dictionary
dict_outin = c(datzen::dictate_outin('dealership','SERVER ROOM'),
                datzen::dictate_outin('big','LAP'),
                datzen::dictate_outin('rim','RAM'))

yankovise(snip,brev=TRUE,suffix="- @2chainz",dict_outin = dict_outin)
}
